+++ test/event_store/subscription_test.exs
@@ -1,11 +1,5 @@
 defmodule Commanded.EventStore.Adapters.Extreme.SubscriptionTest do
-  use Commanded.EventStore.SubscriptionTestCase
-
-  alias Commanded.EventStore.Adapters.Extreme.Storage
-
-  setup do
-    Storage.reset!()
-  end
+  use Commanded.EventStore.SubscriptionTestCase, application: ExtremeApplication
 
   defp event_store_wait(_default \\ nil), do: 5_000
 end